+++ policy_sentry/shared/policy.py
@@ -102,7 +102,7 @@ class ArnActionGroup:
         #     print("Yaml file is missing this block: " + e.args[0])
         #     sys.exit()
         except IndexError:
-            print("IndexError: list index out of range. This is likely due to an ARN in your list equaling ''."
+            print("IndexError: list index out of range. This is likely due to an ARN in your list equaling ''. "
                   "Please evaluate your YML file and try again.")
             sys.exit()